refactored: removed type prefix in model for flags
